renamed contig_methylation -> strand
